fix: remove spec-test-utils
