+++ test/spec/utils.ts
@@ -1,8 +1,14 @@
-import fs, {statSync} from "fs";
-import path from "path";
+import fs from "node:fs";
+import path from "node:path";
+import util from "node:util";
+import stream from "node:stream";
+import tar from "tar";
+import axios from "axios";
 import jsYaml from "js-yaml";
+import {retry} from "@lodestar/utils";
 
 const REPO_ROOT = path.resolve(__dirname, "..", "..");
+export const SPEC_TEST_REPO_URL = "https://github.com/ethereum/consensus-spec-tests";
 
 export const G2_POINT_AT_INFINITY =
   "0xc00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000";
@@ -101,7 +107,7 @@ export function getTestBatch({directory, innerBlsFolder, namedYamlFiles}: TestBa
     const pathSegments = [fullDirPath, functionName];
     if (innerBlsFolder) pathSegments.push("bls");
     const testDirectory = path.resolve(...pathSegments);
-    if (!statSync(testDirectory).isDirectory()) {
+    if (!fs.statSync(testDirectory).isDirectory()) {
       continue;
     }
     const testGroup: TestGroup = {
@@ -117,4 +123,92 @@ export function getTestBatch({directory, innerBlsFolder, namedYamlFiles}: TestBa
 
 export function isBlstError(e: unknown): boolean {
   return (e as Error).message.includes("BLST_ERROR");
+}
+
+const logEmpty = (): void => {};
+
+export type DownloadTestsOptions = {
+  specVersion: string;
+  outputDir: string;
+  /** Root Github URL `https://github.com/ethereum/consensus-spec-tests` */
+  specTestsRepoUrl: string;
+  /** Release files names to download without prefix `["general", "mainnet", "minimal"]` */
+  testsToDownload: string[];
+};
+
+export interface DownloadGenericTestsOptions<TestNames extends string> {
+  specVersion: string;
+  outputDir: string;
+  specTestsRepoUrl: string;
+  testsToDownload: TestNames[];
+}
+
+/**
+ * Download spec tests
+ */
+export async function downloadTests(opts: DownloadTestsOptions, log: (msg: string) => void = logEmpty): Promise<void> {
+  await downloadGenericSpecTests(opts, log);
+}
+
+/**
+ * Generic Github release downloader.
+ * Used by spec tests and SlashingProtectionInterchangeTest
+ */
+export async function downloadGenericSpecTests<TestNames extends string>(
+  {specVersion, specTestsRepoUrl, outputDir, testsToDownload}: DownloadGenericTestsOptions<TestNames>,
+  log: (msg: string) => void = logEmpty
+): Promise<void> {
+  log(`outputDir = ${outputDir}`);
+
+  // Use version.txt as a flag to prevent re-downloading the tests
+  const versionFile = path.join(outputDir, "version.txt");
+  const existingVersion = fs.existsSync(versionFile) && fs.readFileSync(versionFile, "utf8").trim();
+
+  if (existingVersion === specVersion) {
+    return log(`version ${specVersion} already downloaded`);
+  } else {
+    log(`Downloading new version ${specVersion}`);
+  }
+
+  if (fs.existsSync(outputDir)) {
+    log(`Cleaning existing version ${existingVersion} at ${outputDir}`);
+    fs.rmSync(outputDir, {recursive: true, force: true});
+  }
+
+  fs.mkdirSync(outputDir, {recursive: true});
+
+  await Promise.all(
+    testsToDownload.map(async (test) => {
+      const url = `${specTestsRepoUrl ?? SPEC_TEST_REPO_URL}/releases/download/${specVersion}/${test}.tar.gz`;
+
+      await retry(
+        async () => {
+          const {data, headers} = await axios({
+            method: "get",
+            url,
+            responseType: "stream",
+            timeout: 30 * 60 * 1000,
+          });
+
+          const totalSize = headers["content-length"] as string;
+          log(`Downloading ${url} - ${totalSize} bytes`);
+
+          // extract tar into output directory
+          await util.promisify(stream.pipeline)(data, tar.x({cwd: outputDir}));
+
+          log(`Downloaded  ${url}`);
+        },
+        {
+          retries: 3,
+          onRetry: (e, attempt) => {
+            log(`Download attempt ${attempt} for ${url} failed: ${e.message}`);
+          },
+        }
+      );
+
+      // download tar
+    })
+  );
+
+  fs.writeFileSync(versionFile, specVersion);
 }
